back-end/requirements.txt
fastapi[standard]
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlmodel
pydantic-settings